# docs/features/US-012/api-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexus API",
    "description": "Nexus platform REST API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:1000",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/users/{userId}/roles": {
      "get": {
        "tags": [
          "Role Assignment"
        ],
        "summary": "List the active role assignments held by a user within the caller's tenant",
        "operationId": "listRoles",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active role assignments",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed path UUID",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentListResponse"
                }
              }
            }
          },
          "403": {
            "description": "Missing permission or cross-tenant target",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentListResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Role Assignment"
        ],
        "summary": "Assign a role to a user within the caller's tenant",
        "operationId": "assignRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Assignment created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed path or body UUID",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentResponse"
                }
              }
            }
          },
          "403": {
            "description": "Missing permission, cross-tenant target, or caller is not an active TENANT_ADMIN",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentResponse"
                }
              }
            }
          },
          "404": {
            "description": "User or role not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentResponse"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate active assignment",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{userId}/roles/{roleId}": {
      "delete": {
        "tags": [
          "Role Assignment"
        ],
        "summary": "Revoke a role from a user within the caller's tenant",
        "operationId": "revokeRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Role revoked"
          },
          "400": {
            "description": "Malformed path UUID"
          },
          "403": {
            "description": "Missing permission or cross-tenant target"
          },
          "404": {
            "description": "User, role, or active assignment not found"
          },
          "409": {
            "description": "Would revoke the tenant's last active TENANT_ADMIN"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssignRoleRequest": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
          }
        },
        "required": [
          "roleId"
        ]
      },
      "RoleAssignmentResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "roleId": {
            "type": "string"
          },
          "roleName": {
            "type": "string"
          },
          "assignedAt": {
            "type": "string",
            "format": "date-time"
          },
          "assignedBy": {
            "type": "string"
          }
        }
      },
...
      "RoleAssignmentListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleAssignmentResponse"
            }
          }
        }
      },
...
    }
  }
}